# use native nvl expression for SQL NVL and 2 argument COALESCE