--- a/processing/src/main/java/org/apache/druid/segment/virtual/ExpressionPlan.java
+++ b/processing/src/main/java/org/apache/druid/segment/virtual/ExpressionPlan.java
@@ -150,7 +150,7 @@ public class ExpressionPlan
       Parser.validateExpr(applied, applied.analyzeInputs());
       return applied;
     }
-    return expression;
+    return getExpression();
   }
 
   /**
@@ -170,7 +170,7 @@ public class ExpressionPlan
       );
       return Parser.foldUnappliedBindings(expression, analysis, unappliedInputs, accumulatorId);
     }
-    return expression;
+    return getExpression();
   }
 
   /**