--- a/sql/src/main/java/org/apache/druid/sql/calcite/planner/DruidOperatorTable.java
+++ b/sql/src/main/java/org/apache/druid/sql/calcite/planner/DruidOperatorTable.java
@@ -376,6 +376,7 @@ public class DruidOperatorTable implements SqlOperatorTable
                        SqlStdOperatorTable.IS_NOT_FALSE,
                        "> 0"
                    )) // Matches Evals.asBoolean
+                   .add(new DirectOperatorConversion(DruidConvertletTable.NVL_FUNCTION, "nvl"))
                    .add(new BinaryOperatorConversion(SqlStdOperatorTable.MULTIPLY, "*"))
                    .add(new BinaryOperatorConversion(SqlStdOperatorTable.MOD, "%"))
                    .add(new BinaryOperatorConversion(SqlStdOperatorTable.DIVIDE, "/"))