--- a/sql/src/test/java/org/apache/druid/sql/calcite/CalciteMultiValueStringQueryTest.java
+++ b/sql/src/test/java/org/apache/druid/sql/calcite/CalciteMultiValueStringQueryTest.java
@@ -2014,6 +2014,4 @@ public class CalciteMultiValueStringQueryTest extends BaseCalciteQueryTest
 
     );
   }
-
-
 }