… sql/src/test/java/org/apache/druid/sql/calcite/CalciteParameterQueryTest.java
… sql/src/test/java/org/apache/druid/sql/calcite/CalciteQueryTest.java