### Description

Do what the title says. The native `nvl` expression, besides being simpler to evaluate than `case_searched(notnull(...)...)`, it is also vectorized, so this should result in a performance improvement for any queries that would be otherwise vectorizable if not for something like a `NVL(x, 'no x')` or `COALESCE(x, 'no x')`. Multi-argument `COALESCE` still plans to a `CASE` statement, which is not vectorized.

The built-in `COALESCE` operator definition does some super early mandatory rewriting into a `CASE` statement a lot earlier than a convertlet could handle it, so I have updated our handling of `CASE` to look for cases it can rewrite into the native `NVL`. Along with this, I added some additional conversion logic to try to plan what more complicated coalesce/nvl expressions can sometimes plan to, which is an expression filter of the form `case when x is not null then expr1 else expr2 end` instead into something like `expr1 or (expr2 and x is null)` if either of expr1 or expr2 can translate into more optimal native filters.

This should result in a performance improvement in some cases, see benchmarks https://github.com/apache/druid/pull/13897#issuecomment-1461122829

The performance of `NVL` vectorized hasn't really been optimized yet, so there might be some cases where vectorizing isn't always better, but it seems close enough for now. When the remaining conditional expressions are vectorized, my intention is to do them using a `VectorMatch` or something similar, in order to only partially evaluate the remaining conditional arguments as needed, instead of always evaluating all arguments like the current iteration of the vectorized `nvl` expression. At this point, I would expect it to almost always come out ahead of the non-vectorized version.

#### Release note
SQL operators `NVL` and `COALESCE` with 2 arguments now plan to a native `nvl` expression, which supports the vector engine. Multi-argument `COALESCE` still plans into a `case_searched`, which is not vectorized.

<hr>

This PR has:

- [x] been self-reviewed.
- [x] a release note entry in the PR description.
- [x] added comments explaining the "why" and the intent of the code wherever would not be obvious for an unfamiliar reader.
- [x] added unit tests or modified existing tests to cover new code paths, ensuring the threshold for [code coverage](https://github.com/apache/druid/blob/master/dev/code-review/code-coverage.md) is met.
- [x] been tested in a test Druid cluster.