@clintropolis 2023-03-09
redid how i was doing some things and the benchmarks look better all the time now

before:
```
Benchmark              (query)  (rowsPerSegment)  (storageType)  (stringEncoding)  (vectorize)  Mode  Cnt     Score    Error  Units
SqlBenchmark.querySql       28           5000000           mmap              none        false  avgt    5   141.133 ±  3.026  ms/op
SqlBenchmark.querySql       29           5000000           mmap              none        false  avgt    5   877.378 ± 11.754  ms/op
SqlBenchmark.querySql       30           5000000           mmap              none        false  avgt    5   100.516 ±  4.728  ms/op
SqlBenchmark.querySql       31           5000000           mmap              none        false  avgt    5    22.473 ±  0.182  ms/op
```

after:
```
Benchmark              (query)  (rowsPerSegment)  (storageType)  (stringEncoding)  (vectorize)  Mode  Cnt    Score   Error  Units
SqlBenchmark.querySql       28           5000000           mmap              none        false  avgt    5  141.381 ± 1.813  ms/op
SqlBenchmark.querySql       28           5000000           mmap              none        force  avgt    5  103.825 ± 0.657  ms/op
SqlBenchmark.querySql       29           5000000           mmap              none        false  avgt    5  617.000 ± 4.266  ms/op
SqlBenchmark.querySql       29           5000000           mmap              none        force  avgt    5  582.123 ± 4.627  ms/op
SqlBenchmark.querySql       30           5000000           mmap              none        false  avgt    5   95.310 ± 0.528  ms/op
SqlBenchmark.querySql       30           5000000           mmap              none        force  avgt    5   34.568 ± 0.421  ms/op
SqlBenchmark.querySql       31           5000000           mmap              none        false  avgt    5   20.023 ± 0.159  ms/op
SqlBenchmark.querySql       31           5000000           mmap              none        force  avgt    5   18.774 ± 0.678  ms/op
```